por defecto sin decimales

+++ bbdd_elite.sql
@@ -39,8 +39,8 @@ CREATE TABLE public.producto_estacion (
 	id_producto varchar NOT NULL,
 	id_estacion int8 NOT NULL,
 	stock int8 DEFAULT 0 NULL,
-	sellprice float8 NULL,
-	buyprice float8 NULL,
+	sellprice int8 NULL,
+	buyprice int8 NULL,
 	demand int8 DEFAULT 0 NULL,
 	CONSTRAINT producto_estacion_estaciones_fk FOREIGN KEY (id_estacion) REFERENCES public.estaciones(id) ON DELETE SET NULL ON UPDATE SET NULL DEFERRABLE INITIALLY DEFERRED,
 	CONSTRAINT producto_estacion_productos_fk FOREIGN KEY (id_producto) REFERENCES public.productos(id) ON DELETE SET NULL ON UPDATE SET NULL DEFERRABLE INITIALLY DEFERRED
